requirements.txt
streamlit==1.32.0
pytesseract==0.3.10
pdf2image==1.16.3
python-docx==1.0.1
weasyprint==60.1
Pillow==10.2.0